build: Refactor release workflow with 10up deploy action
This commit replaces the manual shell script for creating the plugin zip with the `10up/action-wordpress-plugin-deploy` GitHub Action.

This change simplifies and standardizes the release process, making it more robust. A `.distignore` file has been added to exclude development files from the final zip archive.

The workflow trigger is also updated to run on pushes to any branch except `main`, streamlining the release of new versions.

Signed-off-by: ABCDO-TN <abdelaziz.abbessi@gmail.com>